+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 - Added the MaterialX\:\:GeomPropDef class for geometric input declarations.
 - Added the Document\:\:getGeomAttrValue method.
 - Added the ValueElement\:\:getResolvedValue method.
-- Added support for GCC 8 and Clang 7.
 - Added support for setting search path via environment variable.
+- Added support for GCC 8 and Clang 7.
 
 ### Changed
 - Added callbacks Observer\:\:onCopyContent and Observer\:\:onClearContent, and removed callback Observer::onInitialize.