+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the MaterialX\:\:GeomPropDef class for geometric input declarations.
 - Added the Document\:\:getGeomAttrValue method.
 - Added the ValueElement\:\:getResolvedValue method.
+- Added support for setting search path via environment variable.
 - Added support for GCC 8 and Clang 7.
 
 ### Changed